+++ tests/unit/libraries/Post/TimeTest.php
@@ -1,0 +1,13 @@
+<?php
+declare (strict_types = 1);
+
+namespace GrottoPress\WordPress\Post;
+
+use Codeception\Util\Stub;
+use tad\FunctionMocker\FunctionMocker;
+use GrottoPress\WordPress\Posts\AbstractTestCase;
+use WP_Post;
+
+class TimeTest extends AbstractTestCase
+{
+}