+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.8.0 - 2018-03-22
+
+### Added
+- Unit tests
+
+### Removed
+- Shell script to copy assets from `node_modules` to `assets/vendor`
+
+### Changed
+- Replace WP browser with codeception for tests
+- Update font awesome social icon class names to v5
+
 ## 0.7.2 - 2018-03-01
 
 ### Removed